Refactor redisdb into redis in benchmarking
